add a paranoid check for f.exists() in deleteContents()
git-svn-id: https://wala.svn.sourceforge.net/svnroot/wala/trunk@4059 f5eafffb-2e1d-0410-98e4-8ec43c5233c4

+++ com.ibm.wala.util/src/com/ibm/wala/util/io/FileUtil.java
@@ -106,7 +106,8 @@ public class FileUtil {
 
     for (Iterator<File> it = fl.iterator(); it.hasNext();) {
       File f = (File) it.next();
-      if (!f.isDirectory()) {
+      // add the check for f.exists(), just in case there's a race going on.
+      if (!f.isDirectory() && f.exists()) {
         boolean result = f.delete();
         if (!result) {
           throw new IOException("Failed to delete " + f);
